requirements.txt
python-dotenv==0.17.0
requests==2.25.1
stripe==2.56.0